Use Java6 (!) ServiceLoader to register context handers
To do this, I had to extend Java's HttpHandler interface to add a getContext
method.

Also, add a base handler class that catches any errors in the handler
implementation and returns a 500 error page.  The default behavior is
to silently ignore the exception and carry on as if nothing went wrong.

Update the gold standard for the one test so that it now matches the
new code.