Add minimum bound to memory test
If shell commands that calculate RSS of the process are incorrect
for the environment, one way it fails is the RSS is reported as
zero.

For example, this happens on Alpine because its ps utility outputs
rss in megabytes, not kilobytes.

Adding a lower bound makes the test more likely to catch this.