Add a system logger that logs to console
This was kind of cool how automagically, the server started logging
stuff from the Java libraries.

For example, it pointed out to me that by default, this server
binds to all network adapters (0.0.0.0).  Which is not what I
want.  I added a new host variable to control this in the server,
but did not hook it up yet to a command-line variable.

Also, fix the Makefile to recompile classes when the sources
change.  This stopped working because I used a target name
that matched a directory name.

This commit fixes #3.